refactore: create CoppyLinkButton component
